[refactor] 폴더링 수정 및 서버 통신 파일 변경 (tosspay->data)
