+++ app/src/main/java/com/example/toss_and/data/SrvcInterface.kt
@@ -3,6 +3,8 @@ package com.example.toss_and.data
 import com.example.toss_and.data.model.BrandconDto
 import com.example.toss_and.data.model.ResAssetDto
 import com.example.toss_and.data.model.ResLikeDto
+import com.example.toss_and.data.model.ResponseBrandconDto
+import com.example.toss_and.data.model.ResponseGroupBuyingDto
 import retrofit2.Call
 import retrofit2.http.GET
 import retrofit2.http.Header
@@ -26,4 +28,13 @@ interface SrvcInterface {
         @Header("Authorization") authorization: String = "1",
         @Path("productId") productId: Int
     ): Call<ResLikeDto>
+
+
+    @GET("product")
+    fun getProduct(
+    ): Call<ResponseGroupBuyingDto>
+
+    @GET("product/brand")
+    fun getBrand(
+    ): Call<ResponseBrandconDto>
 }
