+++ app/src/main/res/layout/fragment_toss_pay.xml
@@ -12,10 +12,10 @@
 
         <ImageView
             android:id="@+id/iv_toss_pay_setting"
-            android:layout_width="17dp"
-            android:layout_height="17dp"
+            android:layout_width="24dp"
+            android:layout_height="24dp"
             android:layout_marginTop="6dp"
-            android:layout_marginEnd="20dp"
+            android:layout_marginEnd="10dp"
             android:foregroundGravity="center"
             android:src="@drawable/icn_setting"
             app:layout_constraintEnd_toEndOf="parent"
@@ -162,7 +162,7 @@
                     app:layout_constraintStart_toStartOf="@id/tv_toss_pay_title"
                     app:layout_constraintTop_toBottomOf="@id/tv_category" />
 
-                <com.example.toss_and.presentation.tosspay.CategoryButton
+                <com.example.toss_and.presentation.tosspay.screens.CategoryButton
                     android:id="@+id/item_brandcon1"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
@@ -172,7 +172,7 @@
                     app:layout_constraintTop_toBottomOf="@id/tv_brandcon_detail"
                     app:title="카페" />
 
-                <com.example.toss_and.presentation.tosspay.CategoryButton
+                <com.example.toss_and.presentation.tosspay.screens.CategoryButton
                     android:id="@+id/item_brandcon2"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
@@ -183,7 +183,7 @@
                     app:layout_constraintTop_toBottomOf="@id/tv_brandcon_detail"
                     app:title="상품권" />
 
-                <com.example.toss_and.presentation.tosspay.CategoryButton
+                <com.example.toss_and.presentation.tosspay.screens.CategoryButton
                     android:id="@+id/item_brandcon3"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
@@ -194,7 +194,7 @@
                     app:layout_constraintTop_toBottomOf="@id/tv_brandcon_detail"
                     app:title="치킨" />
 
-                <com.example.toss_and.presentation.tosspay.CategoryButton
+                <com.example.toss_and.presentation.tosspay.screens.CategoryButton
                     android:id="@+id/item_brandcon4"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
@@ -205,7 +205,7 @@
                     app:layout_constraintTop_toBottomOf="@id/tv_brandcon_detail"
                     app:title="피자 · 버거" />
 
-                <com.example.toss_and.presentation.tosspay.CategoryButton
+                <com.example.toss_and.presentation.tosspay.screens.CategoryButton
                     android:id="@+id/item_brandcon5"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
@@ -215,7 +215,7 @@
                     app:layout_constraintTop_toBottomOf="@id/item_brandcon1"
                     app:title="편의점" />
 
-                <com.example.toss_and.presentation.tosspay.CategoryButton
+                <com.example.toss_and.presentation.tosspay.screens.CategoryButton
                     android:id="@+id/item_brandcon6"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
@@ -226,7 +226,7 @@
                     app:layout_constraintTop_toBottomOf="@id/item_brandcon1"
                     app:title="외식" />
 
-                <com.example.toss_and.presentation.tosspay.CategoryButton
+                <com.example.toss_and.presentation.tosspay.screens.CategoryButton
                     android:id="@+id/item_brandcon7"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
@@ -280,7 +280,7 @@
                     app:layout_constraintStart_toStartOf="@id/tv_toss_pay_title"
                     app:layout_constraintTop_toBottomOf="@id/view_grey2" />
 
-                <com.example.toss_and.presentation.tosspay.TosspayShopping
+                <com.example.toss_and.presentation.tosspay.screens.TosspayShopping
                     android:id="@+id/item_shopping1"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
@@ -291,7 +291,7 @@
                     app:layout_constraintTop_toBottomOf="@id/tv_shopping"
                     app:title="캐시백 받을 브랜드 고르기" />
 
-                <com.example.toss_and.presentation.tosspay.TosspayShopping
+                <com.example.toss_and.presentation.tosspay.screens.TosspayShopping
                     android:id="@+id/item_shopping2"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
@@ -301,7 +301,7 @@
                     app:layout_constraintTop_toBottomOf="@id/item_shopping1"
                     app:title="1,000원" />
 
-                <com.example.toss_and.presentation.tosspay.TosspayShopping
+                <com.example.toss_and.presentation.tosspay.screens.TosspayShopping
                     android:id="@+id/item_shopping3"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
@@ -350,7 +350,7 @@
                     app:layout_constraintStart_toStartOf="@id/tv_toss_pay_title"
                     app:layout_constraintTop_toBottomOf="@id/view_grey3" />
 
-                <com.example.toss_and.presentation.tosspay.TosspayBrandRanking
+                <com.example.toss_and.presentation.tosspay.screens.TosspayBrandRanking
                     android:id="@+id/item_ranking1"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
@@ -362,7 +362,7 @@
                     app:title="배달의 민족" />
 
 
-                <com.example.toss_and.presentation.tosspay.TosspayBrandRanking
+                <com.example.toss_and.presentation.tosspay.screens.TosspayBrandRanking
                     android:id="@+id/item_ranking2"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
@@ -373,7 +373,7 @@
                     app:title="무신사 스토어" />
 
 
-                <com.example.toss_and.presentation.tosspay.TosspayBrandRanking
+                <com.example.toss_and.presentation.tosspay.screens.TosspayBrandRanking
                     android:id="@+id/item_ranking3"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
@@ -384,7 +384,7 @@
                     app:title="넥슨" />
 
 
-                <com.example.toss_and.presentation.tosspay.TosspayBrandRanking
+                <com.example.toss_and.presentation.tosspay.screens.TosspayBrandRanking
                     android:id="@+id/item_ranking4"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
@@ -395,7 +395,7 @@
                     app:title="여기어때" />
 
 
-                <com.example.toss_and.presentation.tosspay.TosspayBrandRanking
+                <com.example.toss_and.presentation.tosspay.screens.TosspayBrandRanking
                     android:id="@+id/item_ranking5"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
